+++ dzVents/documentation/history.md
@@ -1,3 +1,8 @@
+[2.4.27]
+- Add attribute protected for devices / scenes and groups
+- Add methods protectionOn and protectionOff for devices / scenes and groups
+- Add functions rpad, lpad, mpad, zpad, numDecimals in utils
+
 [2.4.26]
 - Add Smoke detector